pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "xiaopaw-team"
version = "0.1.0"
description = "Multi-agent team with 4 digital employees (Manager+PM+RD+QA) built on xiaopaw-with-memory framework"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.98.0",
    "lark-oapi>=1.4.0",
    "filelock>=3.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "croniter>=2.0",
    "httpx>=0.27",
    "prometheus-client>=0.20",
    "aiohttp>=3.9",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "respx>=0.21",
    "freezegun>=1.4",
]

[tool.setuptools.packages.find]
include = ["xiaopaw_team*"]
exclude = ["tests*", "docs*", "workspace*", "data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: fast unit tests",
    "integration: integration tests (no LLM, with real filesystem)",
    "e2e: end-to-end tests (real sandbox + real LLM)",
]
